// src/Settings/settings.types.ts
export type Settings = {
  emailInputFieldCSS: string;
  emailInputElementCSS: string;
  emailValidBlockCSS: string;
  emailInvalidBlockCSS: string;
  emailDeleteIconCSS: string;
  emailHolderFocusCSS: string;
  maxLengthEmail: number;
  minLengthEmail: number;
  emailRegEx: RegExp;
  copyPasteSeparator: string;
};